Remove CanvasView and EditableCanvasView protocols and rename CanvasViewClass to CanvasView
