+++ Annotations/Classes/Views/Canvas/CanvasDrawable.swift
@@ -1,11 +1,3 @@
-//
-//  CanvasDrawable.swift
-//  Zappy Arrow Annotation
-//
-//  Created by Mirko on 1/6/19.
-//  Copyright © 2019 Blackbelt Labs. All rights reserved.
-//
-
 import Foundation
 
 public protocol CanvasDrawable: class {
