+++ Example/Annotations/Base.lproj/Main.storyboard
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.Storyboard.XIB" version="3.0" toolsVersion="15505" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" initialViewController="B8D-0N-5wS">
+<document type="com.apple.InterfaceBuilder3.Cocoa.Storyboard.XIB" version="3.0" toolsVersion="17156" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" initialViewController="B8D-0N-5wS">
     <dependencies>
         <deployment identifier="macosx"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="15505"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="17156"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
@@ -713,11 +713,11 @@ CA
                         <rect key="frame" x="0.0" y="0.0" width="511" height="309"/>
                         <autoresizingMask key="autoresizingMask"/>
                         <subviews>
-                            <customView translatesAutoresizingMaskIntoConstraints="NO" id="ebu-rf-uXG" customClass="CanvasViewClass" customModule="Annotations">
+                            <customView translatesAutoresizingMaskIntoConstraints="NO" id="ebu-rf-uXG" customClass="CanvasView" customModule="Annotations">
                                 <rect key="frame" x="0.0" y="0.0" width="511" height="244"/>
                             </customView>
                             <segmentedControl verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="JrP-53-D0T">
-                                <rect key="frame" x="84" y="8" width="419" height="24"/>
+                                <rect key="frame" x="108" y="8" width="395" height="24"/>
                                 <segmentedCell key="cell" borderStyle="border" alignment="left" style="rounded" trackingMode="selectOne" id="Y11-io-MbH">
                                     <font key="font" metaFont="system"/>
                                     <segments>
@@ -848,7 +848,7 @@ CA
                         </constraints>
                     </view>
                     <connections>
-                        <outlet property="canvasView" destination="ebu-rf-uXG" id="Av0-HY-c82"/>
+                        <outlet property="canvasView" destination="ebu-rf-uXG" id="Cxx-yk-2XR"/>
                         <outlet property="pickerViewsStackView" destination="0T1-Hw-OI2" id="PII-RJ-RaC"/>
                         <outlet property="redoButton" destination="Gha-MZ-5cr" id="4nh-bT-OLV"/>
                         <outlet property="undoButton" destination="V4e-VP-LAe" id="EjO-fn-yJJ"/>
